splitting worker connector into standalone and mobile
