+++ workers/unity/Assets/Playground/Scripts/Worker/ClientWorkerConnector.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 namespace Playground
 {
-    public class ClientWorkerConnector : WorkerConnector
+    public class ClientWorkerConnector : StandaloneWorkerConnector
     {
         [SerializeField] private GameObject level;
