+++ workers/unity/Assets/Playground/Scripts/Worker/GameLogicWorkerConnector.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 
 namespace Playground
 {
-    public class GameLogicWorkerConnector : WorkerConnector
+    public class GameLogicWorkerConnector : StandaloneWorkerConnector
     {
         [SerializeField] private GameObject level;
