+++ workers/unity/Packages/com.improbable.gdk.core/Worker/StandaloneWorkerConnector.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e2445a4373f816440afc6f49cc012e66
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
